error if $2 doesn't exist
git-svn-id: svn://svn.code.sf.net/p/nsis/code@26 212acab6-be3b-0410-9dea-997c60f758d6

+++ web-scripts/trunk/updatewiki.sh
@@ -9,9 +9,14 @@
 
 cd /home/groups/n/ns/nsis/htdocs
 
-if [ -f "$2" ]; then
-	tar xzf $2 || exit 1
-	rm -f $2
+if [ -n "$2" ]; then
+	if [ -f "$2" ]; then
+		tar xzf $2 || exit 1
+		rm -f $2
+	else
+		echo "$2 cannot be found!"
+		exit 1
+	fi
 else
 	wget http://osdn.dl.sourceforge.net/sourceforge/wikipedia/mediawiki-$1.tar.gz || exit 1
 	tar xzf mediawiki-$1.tar.gz || exit 1
